moved homPage.js and apiKey.js into src folder
