+++ src/index.js
@@ -1,6 +1,6 @@
 import './style.css';
-import { getPics, postLikes, showLikes } from '../homePage.js';
+import { getPics, postLikes, showLikes } from './homePage.js';
 
 
-console.log(getPics());
+getPics();
 showLikes();
